[WFLY-4584] new messaging-activemq subsystem
* add Maven GAVs for ActiveMQ Artemis dependencies
* update test to use urn:jboss:domain:messaging-activemq:1.0 namespace
* add ActiveMQ Artemis client artifacts to jms client
* add org.apache.activemq.integration.wildfly:activemq-wildfly-integration
artifact
* use service extension to bind with ActiveMQ's transaction manager locator and recovery manager
* add (core) queues addition/removal to JMSOperations

+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/messaging/ArtemisMessagingTestCase.java
@@ -22,32 +22,40 @@
 package org.jboss.as.test.smoke.messaging;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.jboss.as.controller.client.helpers.ClientConstants.OP;
+import static org.jboss.as.controller.client.helpers.ClientConstants.OP_ADDR;
+import static org.jboss.as.controller.client.helpers.ClientConstants.OUTCOME;
+import static org.jboss.as.controller.client.helpers.ClientConstants.READ_RESOURCE_OPERATION;
+import static org.jboss.as.controller.client.helpers.ClientConstants.SUCCESS;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import java.io.IOException;
 import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicReference;
 
-import org.hornetq.api.core.HornetQException;
-import org.hornetq.api.core.TransportConfiguration;
-import org.hornetq.api.core.client.ClientConsumer;
-import org.hornetq.api.core.client.ClientMessage;
-import org.hornetq.api.core.client.ClientProducer;
-import org.hornetq.api.core.client.ClientSession;
-import org.hornetq.api.core.client.ClientSessionFactory;
-import org.hornetq.api.core.client.HornetQClient;
-import org.hornetq.api.core.client.MessageHandler;
-import org.hornetq.core.remoting.impl.invm.InVMConnectorFactory;
+import org.apache.activemq.artemis.api.core.ActiveMQException;
+import org.apache.activemq.artemis.api.core.TransportConfiguration;
+import org.apache.activemq.artemis.api.core.client.ActiveMQClient;
+import org.apache.activemq.artemis.api.core.client.ClientConsumer;
+import org.apache.activemq.artemis.api.core.client.ClientMessage;
+import org.apache.activemq.artemis.api.core.client.ClientProducer;
+import org.apache.activemq.artemis.api.core.client.ClientSession;
+import org.apache.activemq.artemis.api.core.client.ClientSessionFactory;
+import org.apache.activemq.artemis.api.core.client.MessageHandler;
+import org.apache.activemq.artemis.core.remoting.impl.invm.InVMConnectorFactory;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.test.api.ArquillianResource;
+import org.jboss.as.arquillian.container.ManagementClient;
+import org.jboss.dmr.ModelNode;
 import org.jboss.logging.Logger;
-import org.jboss.modules.ModuleClassLoader;
-import org.jboss.modules.ModuleIdentifier;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.After;
+import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -58,34 +66,37 @@ import org.junit.runner.RunWith;
  * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
  */
 @RunWith(Arquillian.class)
-public class MessagingTestCase {
+public class ArtemisMessagingTestCase {
     private static final String QUEUE_EXAMPLE_QUEUE = "queue.exampleQueue";
 
-    static final Logger log = Logger.getLogger(MessagingTestCase.class);
+    static final Logger log = Logger.getLogger(ArtemisMessagingTestCase.class);
 
     private static final String BODY = "msg.body";
 
     private ClientSessionFactory sf;
     private ClientSession session;
 
+
+    @ArquillianResource
+    private ManagementClient managementClient;
+
     @Deployment
     public static JavaArchive createDeployment() throws Exception {
-
         JavaArchive jar = ShrinkWrap.create(JavaArchive.class, "messaging-example.jar");
         jar.addAsManifestResource(new StringAsset("Manifest-Version: 1.0\n" +
-                "Dependencies: org.hornetq, org.jboss.dmr\n"), "MANIFEST.MF");
-        jar.addClass(MessagingTestCase.class);
+                "Dependencies: org.apache.activemq.artemis, org.jboss.dmr, org.jboss.as.controller-client\n"), "MANIFEST.MF");
+        jar.addClass(ArtemisMessagingTestCase.class);
         return jar;
     }
 
     @Before
     public void start() throws Exception {
-        //FIXME Arquillian Alpha bug - it also wants to execute this on the client despite this test being IN_CONTAINER
-        // JFM: I'm not sure this bug is still relevant. I leave it to test it on our CI
-        assertTrue("Test is not run inside the container", isInContainer());
+        System.out.println("managementClient = " + managementClient);
+
+        Assume.assumeTrue("Test is relevant only when the messaging-activemq subsystem is setup", isActiveMQServerFound());
 
         //Not using JNDI so we use the core services directly
-        sf = HornetQClient.createServerLocatorWithoutHA(new TransportConfiguration(InVMConnectorFactory.class.getName())).createSessionFactory();
+        sf = ActiveMQClient.createServerLocatorWithoutHA(new TransportConfiguration(InVMConnectorFactory.class.getName())).createSessionFactory();
         session = sf.createSession();
 
         //Create a queue
@@ -99,13 +110,17 @@ public class MessagingTestCase {
 
     @After
     public void stop() throws Exception {
-        session.close();
+        if (session != null) {
+            session.close();
+        }
 
-        ClientSession coreSession = sf.createSession();
-        coreSession.deleteQueue(QUEUE_EXAMPLE_QUEUE);
-        coreSession.close();
+        if (sf != null) {
+            ClientSession coreSession = sf.createSession();
+            coreSession.deleteQueue(QUEUE_EXAMPLE_QUEUE);
+            coreSession.close();
 
-        sf.close();
+            sf.close();
+        }
     }
 
     @Test
@@ -121,7 +136,7 @@ public class MessagingTestCase {
                     m.acknowledge();
                     message.set(m);
                     latch.countDown();
-                } catch (HornetQException e) {
+                } catch (ActiveMQException e) {
                     e.printStackTrace();
                 }
             }
@@ -144,16 +159,14 @@ public class MessagingTestCase {
         producer.send(message);
     }
 
-    private boolean isInContainer() {
-        ClassLoader cl = this.getClass().getClassLoader();
-        if (cl instanceof ModuleClassLoader == false) {
-            return false;
-        }
-        ModuleClassLoader mcl = (ModuleClassLoader)cl;
-        ModuleIdentifier surefireModule = ModuleIdentifier.fromString("jboss.surefire.module");
-        if (surefireModule.equals(mcl.getModule().getIdentifier())) {
-            return false;
-        }
-        return true;
+    private boolean isActiveMQServerFound() throws IOException {
+        ModelNode readHornetQServer = new ModelNode();
+        readHornetQServer.get(OP).set(READ_RESOURCE_OPERATION);
+        readHornetQServer.get(OP_ADDR).add("subsystem", "messaging-activemq");
+        readHornetQServer.get(OP_ADDR).add("server", "default");
+        System.out.println("readHornetQServer = " + readHornetQServer);
+        ModelNode result = managementClient.getControllerClient().execute(readHornetQServer);
+        System.out.println("result = " + result);
+        return  result.hasDefined(OUTCOME) && SUCCESS.equals(result.get(OUTCOME).asString());
     }
 }
